+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 ChangeLog
 =========
 
+1.0.5 (2024-10-07)
+------------------
+
+- IMPROVEMENT: allow one to use the REST API of the server to control the UI
+- IMPROVEMENT: allow one to use keystrokes to control the UI
+- BUGFIX: fix startup scripts
+- UPDATE: upgrade dependencies
+
 1.0.5 (2024-10-06)
 ------------------